backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
marshmallow==3.20.2
python-dotenv==1.0.0
pytest==7.4.4
PyMySQL==1.1.1
cryptography==46.0.3
